+++ packages/backend-common/src/database/types.ts
@@ -17,11 +17,17 @@
 import knex from 'knex';
 
 /**
- * The PluginDatabaseClientFactory is used to provide a mechanism for backend plugins to obtain database connections
- * for itself.
- *
- * The purpose of this factory is to allow plugins to get isolated data stores so that plugins are discouraged from
- * database integration. Plugins can omit the `database` parameter to get the default plugin database, or
- * provide an identifier that will be used to identify a separate database from the default.
+ * The PluginDatabaseManager manages access to databases that Plugins get.
  */
-export type PluginDatabaseClientFactory = (database?: string) => Promise<knex>;
+export interface PluginDatabaseManager {
+  /**
+   * getClient provides backend plugins database connections for itself.
+   *
+   * The purpose of this method is to allow plugins to get isolated data
+   * stores so that plugins are discouraged from database integration. Plugins
+   * can omit the `database` parameter to get the default plugin database, or
+   * provide an identifier that will be used to identify a separate database
+   * from the default.
+   */
+  getClient(database?: string): Promise<knex>;
+}