+++ packages/create-app/templates/default-app/packages/backend/src/plugins/catalog.ts
@@ -14,12 +14,11 @@ export default async function createPlugin({
   logger,
   config,
   reader,
-  databaseClientFactory,
+  database,
 }: PluginEnvironment) {
   const locationReader = new LocationReaders({ logger, reader, config });
 
-  const db = await DatabaseManager.createDatabase(
-    await databaseClientFactory(),
+  const db = await DatabaseManager.createDatabase(await database.getClient(),
     { logger },
   );
   const entitiesCatalog = new DatabaseEntitiesCatalog(db);