+++ packages/create-app/templates/default-app/packages/backend/src/types.ts
@@ -1,14 +1,14 @@
 import { Logger } from 'winston';
 import { Config } from '@backstage/config';
 import {
-  PluginDatabaseClientFactory,
+  PluginDatabaseManager,
   PluginEndpointDiscovery,
-  UrlReader
+  UrlReader,
 } from '@backstage/backend-common';
 
 export type PluginEnvironment = {
   logger: Logger;
-  databaseClientFactory: PluginDatabaseClientFactory;
+  database: PluginDatabaseManager;
   config: Config;
   reader: UrlReader
   discovery: PluginEndpointDiscovery;